Fixed oembed bug
oembed wanted to rewrite links (e.g. wikipedia URLs) but the links.html
file had a placeholder which generated invalid HTML, resulting in a
malformed like '<a href="#not sure why why this is needed, maybe the
oembed API updated?…' which eats all of the content until the next
double-quote.

This is not a complete fix - I'm not sure whether oembed should be doing
something else here and this generates proper output but doesn't produce
a stored oembed object like the other translations produce.